requirements.txt
asgiref==3.7.2
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==2.0.0
Django==4.2.3
djangorestframework==3.14.0
ecdsa==0.18.0
gunicorn==20.1.0
idna==3.4
psycopg==3.1.9
psycopg-binary==3.1.9
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
